+++ packages/core/package.json
@@ -26,21 +26,20 @@
     "@loadable/component": "^5.10.2",
     "classnames": "^2.2.5",
     "deep-equal": "^1.0.1",
-    "dnd-core": "^2.3.0",
     "element-resize-event": "^2.0.9",
     "lodash.throttle": "^4.1.1",
-    "mousetrap": "^1.6.1",
+    "is-hotkey": "~0.1.6",
     "ramda": "^0.24.1",
     "react-dimensions": "^1.3.0",
-    "react-dnd-cjs": "^9.4.0",
-    "react-dnd-html5-backend-cjs": "^9.4.0",
+    "react-dnd": "^10.0.2",
+    "react-dnd-html5-backend": "^10.0.2",
     "react-draggable": "^3.0.3",
     "react-hotkeys": "^0.10.0",
     "react-redux": "^7.1.1",
     "react-resizable": "^1.9.0",
     "redux": "^4.0.4",
     "redux-thunk": "^2.3.0",
-    "redux-undo": "1.0.0-beta9-7",
+    "redux-undo": "^1.0.0",
     "reselect": "^3.0.0",
     "semver": "^5.3.0",
     "uuid": "^3.0.1"
