+++ packages/core/src/components/Editable/index.tsx
@@ -1,19 +1,15 @@
 import equals from 'ramda/src/equals';
 import * as React from 'react';
-import Editor, { DndBackend } from '../../';
-import { ReduxProvider } from '../../reduxConnect';
+import Editor from '../../Editor';
 import { editable } from '../../selector/editable';
 import { EditorState } from '../../types/editor';
-import DragDropProvider from '../DragDropProvider';
 import HotKeyDecorator from '../HotKey/Decorator';
-import BlurGate from './BlurGate';
 import Inner from './Inner';
 
 export type PropTypes = {
   id: string;
   editor: Editor;
   onChange?: Function;
-  dndBackend?: DndBackend;
 };
 class Editable extends React.PureComponent<PropTypes> {
   unsubscribe: Function;
@@ -70,20 +66,13 @@ class Editable extends React.PureComponent<PropTypes> {
   render() {
     const {
       id,
-      editor: { store, defaultPlugin },
-      dndBackend,
+      editor: { defaultPlugin },
     } = this.props;
 
     return (
-      <ReduxProvider store={store}>
-        <DragDropProvider dndBackend={dndBackend}>
-          <HotKeyDecorator id={id}>
-            <BlurGate>
-              <Inner id={id} defaultPlugin={defaultPlugin} />
-            </BlurGate>
-          </HotKeyDecorator>
-        </DragDropProvider>
-      </ReduxProvider>
+      <HotKeyDecorator id={id}>
+        <Inner id={id} defaultPlugin={defaultPlugin} />
+      </HotKeyDecorator>
     );
   }
 }
