+++ packages/core/src/components/Row/Droppable/index.tsx
@@ -1,27 +1,5 @@
-/*
- * This file is part of ORY Editor.
- *
- * ORY Editor is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Lesser General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * ORY Editor is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Lesser General Public License for more details.
- *
- * You should have received a copy of the GNU Lesser General Public License
- * along with ORY Editor.  If not, see <http://www.gnu.org/licenses/>.
- *
- * @license LGPL-3.0
- * @copyright 2016-2018 Aeneas Rekkas
- * @author Aeneas Rekkas <aeneas+oss@aeneas.io>
- *
- */
-
 import * as React from 'react';
-import { DropTarget as dropTarget } from 'react-dnd-cjs';
+import { ConnectDropTarget, DropTarget as dropTarget } from 'react-dnd';
 import { dragActions } from '../../../actions/cell/drag';
 import { insertActions } from '../../../actions/cell/insert';
 import { connect } from '../../../reduxConnect';
@@ -34,7 +12,7 @@ export type Props = ComponetizedRow & {
   isLayoutMode: boolean;
   isInsertMode: boolean;
   isOverCurrent: boolean;
-  connectDropTarget<T>(e: T): T;
+  connectDropTarget: ConnectDropTarget;
 };
 
 export class Droppable extends React.Component<Props> {
