+++ packages/core/src/service/hover/input.ts
@@ -1,26 +1,4 @@
-/*
- * This file is part of ORY Editor.
- *
- * ORY Editor is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Lesser General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * ORY Editor is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Lesser General Public License for more details.
- *
- * You should have received a copy of the GNU Lesser General Public License
- * along with ORY Editor.  If not, see <http://www.gnu.org/licenses/>.
- *
- * @license LGPL-3.0
- * @copyright 2016-2018 Aeneas Rekkas
- * @author Aeneas Rekkas <aeneas+oss@aeneas.io>
- *
- */
-
-import { DragSourceMonitor } from 'react-dnd-cjs';
+import { DropTargetMonitor } from 'react-dnd';
 import { findDOMNode } from 'react-dom';
 import { ComponetizedCell } from '../../types/editable';
 import { Callbacks, Room, Vector } from '../../types/hover';
@@ -32,7 +10,7 @@ export const computeCurrentDropPosition = (
   actions: Callbacks,
   hover: ComponetizedCell,
   drag: ComponetizedCell,
-  monitor: DragSourceMonitor,
+  monitor: DropTargetMonitor,
   component: React.ReactInstance,
   matrixName: string
 ) => {
@@ -66,7 +44,7 @@ export const computeAndDispatchInsert = (
     ...hover
   }: ComponetizedCell,
   drag: ComponetizedCell,
-  monitor: DragSourceMonitor,
+  monitor: DropTargetMonitor,
   component: React.ReactInstance,
   matrixName: string = '10x10'
 ) =>
@@ -101,7 +79,7 @@ export const computeAndDispatchHover = (
     ...hover
   }: ComponetizedCell,
   drag: ComponetizedCell,
-  monitor: DragSourceMonitor,
+  monitor: DropTargetMonitor,
   component: React.ReactInstance,
   matrixName: string = '10x10'
 ) =>
