+++ packages/editor/src/editor/EditableEditor.tsx
@@ -1,4 +1,4 @@
-import { createEmptyState, Editable, Editor } from '@react-page/core';
+import { createEmptyState, Editable, Editor, Provider } from '@react-page/core';
 import EditorUI from '@react-page/ui';
 import React, { useEffect, useRef } from 'react';
 import StickyWrapper from './StickyWrapper';
@@ -11,34 +11,24 @@ export default ({ plugins, defaultPlugin, value, onChange, dndBackend }) => {
 
   editorRef.current.editorState = value || createEmptyState();
   // updating plugins is not yet supported, so we have an editor ref  that stays the same during lifetime of the editor
-  useEffect(
-    () => {
-      editorRef.current.editor.trigger.editable.update(
-        editorRef.current.editorState
-      );
-    },
-    [editorRef.current.editorState !== value]
-  );
+  useEffect(() => {
+    editorRef.current.editor.trigger.editable.update(
+      editorRef.current.editorState
+    );
+  }, [editorRef.current.editorState !== value]);
 
   const { editor, editorState } = editorRef.current;
 
   return (
-    <StickyWrapper>
-      {stickyNess => (
-        <>
-          <Editable
-            id={editorState.id}
-            editor={editor}
-            onChange={onChange}
-            dndBackend={dndBackend}
-          />
-          <EditorUI
-            editor={editor}
-            stickyNess={stickyNess}
-            dndBackend={dndBackend}
-          />
-        </>
-      )}
-    </StickyWrapper>
+    <Provider editor={editor} dndBackend={dndBackend}>
+      <StickyWrapper>
+        {stickyNess => (
+          <>
+            <Editable id={editorState.id} editor={editor} onChange={onChange} />
+            <EditorUI editor={editor} stickyNess={stickyNess} />
+          </>
+        )}
+      </StickyWrapper>
+    </Provider>
   );
 };
