+++ packages/plugins/content/slate/package.json
@@ -32,7 +32,7 @@
     "uniforms-bridge-json-schema": "~2.5.0",
     "uniforms-material": "~2.5.0",
     "deep-rename-keys": "^0.2.1",
-    "is-hotkey": "0.1.4",
+    "is-hotkey": "^0.1.6",
     "jsdom": "^15.1.1",
     "lodash.debounce": "^4.0.8",
     "lodash.throttle": "^4.1.1",
@@ -40,7 +40,7 @@
     "lodash.isobject": "^3.0.2",
     "lodash.flatten": "^4.4.0",
     "react-portal": "4.1.5",
-    "redux-undo": "1.0.0-beta9-7",
+    "redux-undo": "^1.0.0",
     "slate": "^0.57.0",
     "slate-react": "^0.57.0",
     "slate-hyperscript": "^0.57.1",
