+++ packages/ui/package.json
@@ -34,8 +34,6 @@
     "lodash.throttle": "^4.1.1",
     "rc-tooltip": "^3.4.3",
     "react-color": "^2.14.1",
-    "react-dnd-cjs": "^9.4.0",
-    "react-dnd-html5-backend-cjs": "^9.4.0",
     "reselect": "^3.0.0"
   },
   "devDependencies": {
