+++ packages/ui/src/EditorUI/index.tsx
@@ -1,27 +1,23 @@
-import { DndBackend, Editor } from '@react-page/core';
 import React from 'react';
 import DisplayModeToggle, { StickyNess } from '../DisplayModeToggle/index';
-import Provider from '../Provider';
+import ThemeProvider from '../ThemeProvider';
 import Toolbar from '../Toolbar/index';
 import Trash from '../Trash/index';
 
-export default ({
-  editor,
-  stickyNess = {
-    shouldStickToTop: false,
-    shouldStickToBottom: false,
-    rightOffset: 0,
-  },
-  dndBackend,
-}: {
-  // tslint:disable-next-line:no-any
-  editor: Editor;
-  stickyNess?: StickyNess;
-  dndBackend?: DndBackend;
-}) => (
-  <Provider editor={editor} dndBackend={dndBackend}>
-    <Trash />
-    <DisplayModeToggle stickyNess={stickyNess} />
-    <Toolbar />
-  </Provider>
+export default React.memo(
+  ({
+    stickyNess = {
+      shouldStickToTop: false,
+      shouldStickToBottom: false,
+      rightOffset: 0,
+    },
+  }: {
+    stickyNess?: StickyNess;
+  }) => (
+    <ThemeProvider>
+      <Trash />
+      <DisplayModeToggle stickyNess={stickyNess} />
+      <Toolbar />
+    </ThemeProvider>
+  )
 );
